chore(vscode): use rolldown for bundling
